try fixing type of data values

+++ tests/utils_test.py
@@ -391,7 +391,9 @@ def blockmodel_tensormesh_equivalency(ws, origin, x, y, z, ignore_origin=False):
         {
             "DataValues": {
                 "association": "CELL",
-                "values": np.random.randint(0, 100, size=len(block_model.centroids)),
+                "values": np.random.randint(
+                    0, 100, size=len(block_model.centroids)
+                ).astype(np.int32),
             }
         }
     )
@@ -452,7 +454,9 @@ def test_block_model_2_tensor_negativeh(tmp_path):
         {
             "DataValues": {
                 "association": "CELL",
-                "values": np.random.randint(0, 100, size=len(block_model.centroids)),
+                "values": np.random.randint(
+                    0, 100, size=len(block_model.centroids)
+                ).astype(np.int32),
             }
         }
     )
